CLOUDP-350407: Public interface tweaks for translate
