+++ internal/autogen/translate/crds/crds.go
@@ -23,14 +23,14 @@ import (
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 )
 
-// AssertMajorVersion checks the given majorVersion exists for the given kind
-// and CRD version
+// AssertMajorVersion checks the given SDK majorVersion exists for the given
+// Kubernetes kind and CRD version
 func AssertMajorVersion(specVersion *apiextensionsv1.CustomResourceDefinitionVersion, kind string, majorVersion string) error {
-	props, err := GetOpenAPIProperties(kind, specVersion)
+	props, err := getOpenAPIProperties(kind, specVersion)
 	if err != nil {
 		return fmt.Errorf("failed to enumerate CRD schema properties: %w", err)
 	}
-	specProps, err := GetSpecPropertiesFor(kind, props, "spec")
+	specProps, err := getSpecPropertiesFor(kind, props, "spec")
 	if err != nil {
 		return fmt.Errorf("failed to enumerate CRD spec properties: %w", err)
 	}
@@ -41,7 +41,9 @@ func AssertMajorVersion(specVersion *apiextensionsv1.CustomResourceDefinitionVer
 	return nil
 }
 
-// CompileCRDSchema compiles the given JSON schema properties
+// CompileCRDSchema compiles the given JSON properties as a type schema.
+// Enables validating whether or not some JSON or unstructured data comforms
+// with such type schema.
 func CompileCRDSchema(openAPISchema *apiextensionsv1.JSONSchemaProps) (*jsonschema.Schema, error) {
 	schemaBytes, err := json.Marshal(openAPISchema)
 	if err != nil {
@@ -58,8 +60,7 @@ func CompileCRDSchema(openAPISchema *apiextensionsv1.JSONSchemaProps) (*jsonsche
 	return schema, nil
 }
 
-// SelectVersion returns the version from the CRD spec that matches the given
-// version string
+// SelectVersion extracts the CRD version definition matching the given version.
 func SelectVersion(spec *apiextensionsv1.CustomResourceDefinitionSpec, version string) *apiextensionsv1.CustomResourceDefinitionVersion {
 	if len(spec.Versions) == 0 {
 		return nil
@@ -75,9 +76,9 @@ func SelectVersion(spec *apiextensionsv1.CustomResourceDefinitionSpec, version s
 	return nil
 }
 
-// GetOpenAPIProperties digs up the schema properties of a given kind on a given
-// CRD version
-func GetOpenAPIProperties(kind string, version *apiextensionsv1.CustomResourceDefinitionVersion) (map[string]apiextensionsv1.JSONSchemaProps, error) {
+// getOpenAPIProperties extracts the schema properties of a given CRD version.
+// The kind is passed as a reference of the object this version is from.
+func getOpenAPIProperties(kind string, version *apiextensionsv1.CustomResourceDefinitionVersion) (map[string]apiextensionsv1.JSONSchemaProps, error) {
 	if version == nil {
 		return nil, fmt.Errorf("missing version (nil) from %v spec", kind)
 	}
@@ -93,9 +94,9 @@ func GetOpenAPIProperties(kind string, version *apiextensionsv1.CustomResourceDe
 	return version.Schema.OpenAPIV3Schema.Properties, nil
 }
 
-// GetSpecPropertiesFor takes the properties value of a given field of a kind's
+// getSpecPropertiesFor takes the properties value of a given field of a kind's
 // properties set
-func GetSpecPropertiesFor(kind string, props map[string]apiextensionsv1.JSONSchemaProps, field string) (map[string]apiextensionsv1.JSONSchemaProps, error) {
+func getSpecPropertiesFor(kind string, props map[string]apiextensionsv1.JSONSchemaProps, field string) (map[string]apiextensionsv1.JSONSchemaProps, error) {
 	prop, ok := props[field]
 	if !ok {
 		return nil, fmt.Errorf("kind %q spec is missing field %q on", kind, field)
