+++ internal/autogen/translate/testdata/samples/v1/backupcompliancepolicy.go
