+++ internal/autogen/translate/testdata/samples/v1/backupschedule.go
