+++ internal/autogen/translate/testdata/samples/v1/cluster.go
