+++ internal/autogen/translate/testdata/samples/v1/customrole.go
