+++ internal/autogen/translate/testdata/samples/v1/databaseuser.go
