+++ internal/autogen/translate/testdata/samples/v1/datafederation.go
