+++ internal/autogen/translate/testdata/samples/v1/flexcluster.go
