+++ internal/autogen/translate/testdata/samples/v1/group.go
