+++ internal/autogen/translate/testdata/samples/v1/groupalertsconfig.go
