+++ internal/autogen/translate/testdata/samples/v1/networkpeeringconnection.go
