+++ internal/autogen/translate/testdata/samples/v1/networkpermissionentries.go
