+++ internal/autogen/translate/testdata/samples/v1/organization.go
