+++ internal/autogen/translate/testdata/samples/v1/organizationsetting.go
