+++ internal/autogen/translate/testdata/samples/v1/scheme.go
