+++ internal/autogen/translate/testdata/samples/v1/searchindex.go
