+++ internal/autogen/translate/testdata/samples/v1/team.go
