+++ internal/autogen/translate/testdata/samples/v1/thirdpartyintegration.go
