+++ internal/autogen/translate/testdata/samples/v1/zz_generated.deepcopy.go
