+++ internal/autogen/translate/unstructured/unstructured_test.go
@@ -16,7 +16,6 @@
 package unstructured_test
 
 import (
-	"sort"
 	"testing"
 	"time"
 
@@ -96,171 +95,4 @@ func TestToAndFromAndCopyUnstructured(t *testing.T) {
 	result := testStruct{}
 	require.NoError(t, unstructured.FromUnstructured(&result, sampleMap))
 	assert.Equal(t, sample, result)
-}
-
-func TestCreateAndAccessField(t *testing.T) {
-	sampleMap, err := unstructured.ToUnstructured(sample)
-	require.NoError(t, err)
-	want := "value"
-	require.NoError(t, unstructured.CreateField(sampleMap, "value", "SubObj", "field"))
-	got, err := unstructured.GetField[string](sampleMap, "SubObj", "field")
-	require.NoError(t, err)
-	assert.Equal(t, want, got)
-}
-
-func TestAsPath(t *testing.T) {
-	for _, tc := range []struct {
-		title string
-		input string
-		want  []string
-	}{
-		{
-			title: "nothing",
-			input: "",
-			want:  []string{""},
-		},
-		{
-			title: "single",
-			input: "dir",
-			want:  []string{"dir"},
-		},
-		{
-			title: "double",
-			input: "dir0.dir1",
-			want:  []string{"dir0", "dir1"},
-		},
-		{
-			title: "rooted double",
-			input: ".dir0.dir1",
-			want:  []string{"dir0", "dir1"},
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			assert.Equal(t, tc.want, unstructured.AsPath(tc.input))
-		})
-	}
-}
-
-func TestBase(t *testing.T) {
-	for _, tc := range []struct {
-		title string
-		input []string
-		want  string
-	}{
-		{
-			title: "none",
-			input: []string{},
-			want:  "",
-		},
-		{
-			title: "single",
-			input: []string{"dir"},
-			want:  "dir",
-		},
-		{
-			title: "double",
-			input: []string{"dir0", "dir1"},
-			want:  "dir1",
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			assert.Equal(t, tc.want, unstructured.Base(tc.input))
-		})
-	}
-}
-
-func TestSkipKeysAndFieldsOf(t *testing.T) {
-	for _, tc := range []struct {
-		title string
-		skips []string
-		want  []string
-	}{
-		{
-			title: "no skips",
-			skips: []string{},
-			want:  []string{"ID", "YesNo", "Text", "Data", "Timestamp", "SubStruct"},
-		},
-		{
-			title: "no matchs",
-			skips: []string{"bleh", "ah"},
-			want:  []string{"ID", "YesNo", "Text", "Data", "Timestamp", "SubStruct"},
-		},
-		{
-			title: "last missing",
-			skips: []string{"SubStruct"},
-			want:  []string{"ID", "YesNo", "Text", "Data", "Timestamp"},
-		},
-		{
-			title: "all gone missing",
-			skips: []string{"ID", "YesNo", "Text", "Data", "Timestamp", "SubStruct"},
-			want:  []string{},
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			sampleMap, err := unstructured.ToUnstructured(sample)
-			require.NoError(t, err)
-			sort.Strings(tc.want)
-			filtered := unstructured.SkipKeys(sampleMap, tc.skips...)
-			fields := unstructured.FieldsOf(filtered)
-			sort.Strings(fields)
-			assert.Equal(t, tc.want, fields)
-		})
-	}
-}
-
-func TestAccessOrCreateField(t *testing.T) {
-	for _, tc := range []struct {
-		title        string
-		obj          map[string]any
-		path         []string
-		defaultValue any
-		want         any
-		wantErr      string
-	}{
-		{
-			title:        "create missing",
-			obj:          map[string]any{},
-			path:         []string{"deep", "field"},
-			defaultValue: "some string",
-			want:         "some string",
-		},
-		{
-			title: "read existing",
-			obj: map[string]any{
-				"deep": map[string]any{
-					"field": "other string",
-				},
-			},
-			path:         []string{"deep", "field"},
-			defaultValue: "some string",
-			want:         "other string",
-		},
-		{
-			title:        "nil object",
-			obj:          nil,
-			path:         []string{"deep", "field"},
-			defaultValue: "some string",
-			wantErr:      "nil object",
-		},
-		{
-			title: "wrong path type",
-			obj: map[string]any{
-				"deep": "an string",
-			},
-			path:         []string{"deep", "field"},
-			defaultValue: "some string",
-			wantErr:      "intermediate path [deep] exists but is of type string",
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			got, err := unstructured.GetOrCreateField(tc.obj, tc.defaultValue, tc.path...)
-			if tc.wantErr != "" {
-				require.Nil(t, got)
-				assert.ErrorContains(t, err, tc.wantErr)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tc.want, got)
-			}
-		})
-	}
 }
